Use --output-target to define the object's output format
A recent binutils post version 2.45 change [0] made more strict the usage
of objcopy's target options. In prior versions, the --target option could
be used to specify either an input or output binary object target.

But now, users must use --input-target or --output-target if the input and
output object format differs. The --target option shall only be used when
both input and output formats are the same.

Not doing that, leads to objcopy complaining that the file format is not
recognized:

    /usr/bin/objcopy: efi/fwup.so: file format not recognized

[0]: https://sourceware.org/git/?p=binutils-gdb.git;a=commit;h=5e83077d552e